Factor out TableMatcher
